Factor out a seprate OutgoingRequestProcessor
